+++ src/components/Dropdown/index.ts
@@ -1,2 +1,2 @@
 export { Dropdown } from './Dropdown';
-export { Props as DropdownProps } from './Dropdown.types';
+export { Props as DropdownProps, DropdownListProps } from './Dropdown.types';
